+++ bin-src/tasks/gitInfo.ts
@@ -23,7 +23,7 @@ import { Context, Task } from '../types';
 import { getChangedFilesWithReplacement } from '../git/getChangedFilesWithReplacement';
 import replacedBuild from '../ui/messages/info/replacedBuild';
 import forceRebuildHint from '../ui/messages/info/forceRebuildHint';
-import { getChangedPackageManifests } from '../lib/comparePackageManifests';
+import { getChangedPackageManifests } from '../lib/getChangedPackageManifests';
 
 const SkipBuildMutation = `
   mutation SkipBuildMutation($commit: String!, $branch: String, $slug: String) {
